removed multithreading for extensions because of performence issues.
other small changes
