+++ GNet/Model/Layers/Dropout.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using GNet.GlobalRandom;
 
 namespace GNet.Layers
 {
