+++ GNet/Program.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 
 namespace GNet
 {
@@ -7,16 +8,16 @@ namespace GNet
     {
         private static void Main()
         {
-            ExtensionManager.Initialize(new Extensions.CPU());
-
             var net = new Network
             (
                 new Layers.Dense(new Shape(10), new Activations.Identity(), new Initializers.Zero(), new Initializers.Zero()),
-                new Layers.Dense(new Shape(5, 2), new Activations.Tanh(), new Initializers.LeCunNormal(), new Initializers.Zero()),
+                new Layers.Dense(new Shape(5, 5), new Activations.Tanh(), new Initializers.LeCunNormal(), new Initializers.Zero()),
+                new Layers.Dense(new Shape(5, 5), new Activations.Tanh(), new Initializers.LeCunNormal(), new Initializers.Zero()),
+                new Layers.Dense(new Shape(5, 5), new Activations.Tanh(), new Initializers.LeCunNormal(), new Initializers.Zero()),
                 new Layers.Dense(new Shape(1), new Activations.Sigmoid(), new Initializers.Normal(), new Initializers.Zero())
             );
 
-            var log = new Log(net);
+            var log = new Logger(net);
 
             net.Initialize();
 
@@ -24,7 +25,15 @@ namespace GNet
             Dataset trainingDataset = datasetGenerator.Generate(2000);
             Dataset validationDataset = datasetGenerator.Generate(1000);
 
-            net.Train(trainingDataset, new Losses.MSE(), new Optimizers.NestrovMomentum(), 20, 1000, 0.01, validationDataset, new OutTransformers.Losses.BinaryRoundLoss());
+            Stopwatch sw = new Stopwatch();
+
+            sw.Start();
+
+            net.Train(trainingDataset, new Losses.MSE(), new Optimizers.NestrovMomentum(), 20, 300, 0.00001, validationDataset, new OutTransformers.Losses.BinaryRoundLoss());
+
+            sw.Stop();
+
+            Console.WriteLine(sw.ElapsedMilliseconds);
 
             Console.ReadKey();
         }
